Follow suggestions from @Bahex
Value is now always evaluated once if a closure, so there's no need for either `--no-eval` or `--eval-once`. The closure is also always run without any input.

Because of that, the column code can be deduplicated again, and the default value can be calculated once at the start of the function, so cells won't store `Value::Error` values anymore.