Add lazy closure evaluation to records and list, with cached values
This will only ever run the closure one time, when needed, then cache the values, so long operations are only ever run once, or not at all if not necessary.

Another change is allowing multiple column names, and erroring when the user gives a column name, whereas previously the command would always return the pipeline input.

Also changed is the function signature, since `default` never supported ByteStream input, and this doesn't add it either. This change and the added error can be reverted if desired.