Restrict `config.show_banner` to valid options
This option was still accepting full `Value`s and storing them as such
in the `Config`, at runtime any value not matching `false` or `"short"`
would be treated as `true`.
This PR introduces a `BannerKind` enum to properly represent them and
provide an error message for invalid values.
Makes the config smaller but adds minimal validation overhead.