refactor: switch back to simpler signature, spoof string representation
Only reason for including SyntaxShape::Closure in the signature was to
make it clear in help texts that closures can be used as the argument.

Simplifying the signature simplifies the parsing, so using a
RowCondition on its own and instead making it always look like a union
with `closure(any)` solves the issue without any changes in the parser.

Also, RowCondition will always accept closures anyway, so its textual
representation should always indicate that without the need for OneOf.